refactor : NotificationController.java
